// API.ts
export type ModelTournamentConnection = {
  __typename: "ModelTournamentConnection",
  items:  Array<Tournament | null >,
  nextToken?: string | null,
};
// ...
export type ModelCategoryConditionInput = {
  alias?: ModelStringInput | null,
  and?: Array< ModelCategoryConditionInput | null > | null,
  createdAt?: ModelStringInput | null,
  description?: ModelStringInput | null,
  not?: ModelCategoryConditionInput | null,
  or?: Array< ModelCategoryConditionInput | null > | null,
  updatedAt?: ModelStringInput | null,
};

export type CreateCategoryInput = {
  alias: string,
  description: string,
  id?: string | null,
};
// ...
export type CreateCategoryMutationVariables = {
  condition?: ModelCategoryConditionInput | null,
  input: CreateCategoryInput,
};

export type CreateCategoryMutation = {
  createCategory?:  {
    __typename: "Category",
    alias: string,
    createdAt: string,
    description: string,
    id: string,
    tournament?:  {
      __typename: "ModelTournamentConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type CreateTicketMutationVariables = {
  condition?: ModelTicketConditionInput | null,
  input: CreateTicketInput,
};

export type CreateTicketMutation = {
  createTicket?:  {
    __typename: "Ticket",
    createdAt: string,
    id: string,
    price: number,
    tournament?:  {
      __typename: "Tournament",
      categoryId: string,
      createdAt: string,
      description: string,
      endDate: string,
      gameName: string,
      id: string,
      name: string,
      prize: number,
      registrationPrice: number,
      startDate: string,
      updatedAt: string,
      userId: string,
    } | null,
    tournamentId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      id: string,
      name: string,
      profileOwner?: string | null,
      tierId?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type CreateTierMutationVariables = {
  condition?: ModelTierConditionInput | null,
  input: CreateTierInput,
};

export type CreateTierMutation = {
  createTier?:  {
    __typename: "Tier",
    alias?: string | null,
    createdAt?: string | null,
    description?: string | null,
    id: string,
    tournamentsLimit?: number | null,
    updatedAt?: string | null,
    users?:  {
      __typename: "ModelUserConnection",
      nextToken?: string | null,
    } | null,
    value?: number | null,
  } | null,
};

export type CreateTournamentMutationVariables = {
  condition?: ModelTournamentConditionInput | null,
  input: CreateTournamentInput,
};

export type CreateTournamentMutation = {
  createTournament?:  {
    __typename: "Tournament",
    category?:  {
      __typename: "Category",
      alias: string,
      createdAt: string,
      description: string,
      id: string,
      updatedAt: string,
    } | null,
    categoryId: string,
    createdAt: string,
    description: string,
    endDate: string,
    gameName: string,
    id: string,
    name: string,
    prize: number,
    registrationPrice: number,
    startDate: string,
    tickets?:  {
      __typename: "ModelTicketConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      id: string,
      name: string,
      profileOwner?: string | null,
      tierId?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type CreateUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: CreateUserInput,
};

export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    createdAt: string,
    email: string,
    id: string,
    name: string,
    profileOwner?: string | null,
    tickets?:  {
      __typename: "ModelTicketConnection",
      nextToken?: string | null,
    } | null,
    tier?:  {
      __typename: "Tier",
      alias?: string | null,
      createdAt?: string | null,
      description?: string | null,
      id: string,
      tournamentsLimit?: number | null,
      updatedAt?: string | null,
      value?: number | null,
    } | null,
    tierId?: string | null,
    tournaments?:  {
      __typename: "ModelTournamentConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type DeleteCategoryMutationVariables = {
  condition?: ModelCategoryConditionInput | null,
  input: DeleteCategoryInput,
};

export type DeleteCategoryMutation = {
  deleteCategory?:  {
    __typename: "Category",
    alias: string,
    createdAt: string,
    description: string,
    id: string,
    tournament?:  {
      __typename: "ModelTournamentConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type DeleteTicketMutationVariables = {
  condition?: ModelTicketConditionInput | null,
  input: DeleteTicketInput,
};

export type DeleteTicketMutation = {
  deleteTicket?:  {
    __typename: "Ticket",
    createdAt: string,
    id: string,
    price: number,
    tournament?:  {
      __typename: "Tournament",
      categoryId: string,
      createdAt: string,
      description: string,
      endDate: string,
      gameName: string,
      id: string,
      name: string,
      prize: number,
      registrationPrice: number,
      startDate: string,
      updatedAt: string,
      userId: string,
    } | null,
    tournamentId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      id: string,
      name: string,
      profileOwner?: string | null,
      tierId?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type DeleteTierMutationVariables = {
  condition?: ModelTierConditionInput | null,
  input: DeleteTierInput,
};

export type DeleteTierMutation = {
  deleteTier?:  {
    __typename: "Tier",
    alias?: string | null,
    createdAt?: string | null,
    description?: string | null,
    id: string,
    tournamentsLimit?: number | null,
    updatedAt?: string | null,
    users?:  {
      __typename: "ModelUserConnection",
      nextToken?: string | null,
    } | null,
    value?: number | null,
  } | null,
};

export type DeleteTournamentMutationVariables = {
  condition?: ModelTournamentConditionInput | null,
  input: DeleteTournamentInput,
};

export type DeleteTournamentMutation = {
  deleteTournament?:  {
    __typename: "Tournament",
    category?:  {
      __typename: "Category",
      alias: string,
      createdAt: string,
      description: string,
      id: string,
      updatedAt: string,
    } | null,
    categoryId: string,
    createdAt: string,
    description: string,
    endDate: string,
    gameName: string,
    id: string,
    name: string,
    prize: number,
    registrationPrice: number,
    startDate: string,
    tickets?:  {
      __typename: "ModelTicketConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      id: string,
      name: string,
      profileOwner?: string | null,
      tierId?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type DeleteUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: DeleteUserInput,
};

export type DeleteUserMutation = {
  deleteUser?:  {
    __typename: "User",
    createdAt: string,
    email: string,
    id: string,
    name: string,
    profileOwner?: string | null,
    tickets?:  {
      __typename: "ModelTicketConnection",
      nextToken?: string | null,
    } | null,
    tier?:  {
      __typename: "Tier",
      alias?: string | null,
      createdAt?: string | null,
      description?: string | null,
      id: string,
      tournamentsLimit?: number | null,
      updatedAt?: string | null,
      value?: number | null,
    } | null,
    tierId?: string | null,
    tournaments?:  {
      __typename: "ModelTournamentConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type UpdateCategoryMutationVariables = {
  condition?: ModelCategoryConditionInput | null,
  input: UpdateCategoryInput,
};

export type UpdateCategoryMutation = {
  updateCategory?:  {
    __typename: "Category",
    alias: string,
    createdAt: string,
    description: string,
    id: string,
    tournament?:  {
      __typename: "ModelTournamentConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type UpdateTicketMutationVariables = {
  condition?: ModelTicketConditionInput | null,
  input: UpdateTicketInput,
};

export type UpdateTicketMutation = {
  updateTicket?:  {
    __typename: "Ticket",
    createdAt: string,
    id: string,
    price: number,
    tournament?:  {
      __typename: "Tournament",
      categoryId: string,
      createdAt: string,
      description: string,
      endDate: string,
      gameName: string,
      id: string,
      name: string,
      prize: number,
      registrationPrice: number,
      startDate: string,
      updatedAt: string,
      userId: string,
    } | null,
    tournamentId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      id: string,
      name: string,
      profileOwner?: string | null,
      tierId?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type UpdateTierMutationVariables = {
  condition?: ModelTierConditionInput | null,
  input: UpdateTierInput,
};

export type UpdateTierMutation = {
  updateTier?:  {
    __typename: "Tier",
    alias?: string | null,
    createdAt?: string | null,
    description?: string | null,
    id: string,
    tournamentsLimit?: number | null,
    updatedAt?: string | null,
    users?:  {
      __typename: "ModelUserConnection",
      nextToken?: string | null,
    } | null,
    value?: number | null,
  } | null,
};

export type UpdateTournamentMutationVariables = {
  condition?: ModelTournamentConditionInput | null,
  input: UpdateTournamentInput,
};

export type UpdateTournamentMutation = {
  updateTournament?:  {
    __typename: "Tournament",
    category?:  {
      __typename: "Category",
      alias: string,
      createdAt: string,
      description: string,
      id: string,
      updatedAt: string,
    } | null,
    categoryId: string,
    createdAt: string,
    description: string,
    endDate: string,
    gameName: string,
    id: string,
    name: string,
    prize: number,
    registrationPrice: number,
    startDate: string,
    tickets?:  {
      __typename: "ModelTicketConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      id: string,
      name: string,
      profileOwner?: string | null,
      tierId?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type UpdateUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: UpdateUserInput,
};

export type UpdateUserMutation = {
  updateUser?:  {
    __typename: "User",
    createdAt: string,
    email: string,
    id: string,
    name: string,
    profileOwner?: string | null,
    tickets?:  {
      __typename: "ModelTicketConnection",
      nextToken?: string | null,
    } | null,
    tier?:  {
      __typename: "Tier",
      alias?: string | null,
      createdAt?: string | null,
      description?: string | null,
      id: string,
      tournamentsLimit?: number | null,
      updatedAt?: string | null,
      value?: number | null,
    } | null,
    tierId?: string | null,
    tournaments?:  {
      __typename: "ModelTournamentConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};